refactor: importパスを分割後のlib/actionsに更新し、不要ファイル・デッドコードを削除: actions分割を完了させるため #8
